payroll_advanced/requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
pandas==2.1.4
openpyxl==3.1.2
fpdf==1.7.2
python-dotenv==1.0.0